requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.31.0
flask-cors==4.0.0
PyJWT==2.8.0
Flask-Limiter==3.5.0
limits==3.9.0
